Api: Remove accidental second derived class
